eec_hists, ehijing-test modded to support eec calculation

+++ energy_correlator/calc_proj_energy.C
@@ -6,7 +6,7 @@ R__LOAD_LIBRARY(libeicsmear);
 const Double_t Mp(0.9383);
 const Double_t Me(0.511E-3);
 
-void calc_proj_energy(float proj_lab_e = 10, float targ_lab_e = 100, float proj_mass = Me, float targ_mass = Mp)
+void calc_proj_energy(double proj_lab_e = 10, double targ_lab_e = 100, double proj_mass = Me, double targ_mass = Mp)
 {
   // proj_lab is desired projectile energy in lab frame
   // trag_lab is desired target energy in lab frame
